pyproject.toml
[project]
name = "llm-zoomcamp-project-2"
version = "0.1.0"

description = "Harvests Python official documentation and PyPI package documentation for RAG-based LLM applications."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
	"requests>=2.31.0",
	"beautifulsoup4>=4.12.2",
	"sphinx>=7.2.6",
	"langchain>=0.1.20",
	"chromadb>=0.4.22",
	"openai>=1.25.0",
	"fastapi>=0.110.0",
	"uvicorn>=0.37.0",
	"langchain-text-splitters>=0.0.1",
	"rank_bm25>=0.2.2",
	"langchain_openai>=0.0.1",
	"langchain_community>=0.0.21",
	"sentence-transformers>=2.2.0",
	"streamlit>=1.28.0",
	"python-dotenv>=1.0.0",
	"prometheus-client>=0.19.0"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
